requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ML Models
lightgbm>=4.0.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment Tracking
wandb>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
joblib>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
